% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river-get.R
\name{attr_values}
\alias{attr_values}
\title{Get values for all attributes from a model instance}
\usage{
attr_values(instance, attr, hasModel)
}
\arguments{
\item{instance}{A river instance}

\item{attr}{An attribute string}

\item{hasModel}{Boolean}
}
\value{
A single row tibble
}
\description{
Get values for all attributes from a model instance
}
