% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river-get.R
\name{get_instances}
\alias{get_instances}
\title{Get instances of a model}
\usage{
get_instances(model, limit = 100, offset = 0)
}
\arguments{
\item{model}{A string that is a model name}

\item{limit}{Numeric. Default is 100}

\item{offset}{Numeric. Default is 0}
}
\value{
An object of class "response"
}
\description{
Get instances of a model
}
\examples{
\dontrun{
resp <- river_data("model_name")
}
}
