% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river-get.R
\name{instance_data}
\alias{instance_data}
\title{Parse data for all attributes from all instances}
\usage{
instance_data(resp)
}
\arguments{
\item{resp}{A response from \code{get_instances}}
}
\value{
A tibble
}
\description{
Parse data for all attributes from all instances
}
\examples{
\dontrun{
resp <- river_data("model_name")
data_output <- instance_data(resp)
}
}
